requirements.txt
streamlit
mysql-connector-python
sqlalchemy
pandas
plotly
ollama
ollama pull mistral
ollama pull llama3
ollama pull phi3